requirements.txt
selenium
requests
beautifulsoup4
webdriver_manager
python-telegram-bot
gunicorn